Add ECS and new Vulkan rendering systems

+++ BansheeEngine/BansheeEngine/source/graphics/Vulkan/VulkanFence.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "foundation/Platform.h"
 #include <vector>
 
 typedef struct VkDevice_T* VkDevice;
@@ -11,15 +10,15 @@ namespace Banshee
 	class VulkanFence
 	{
 	public:
-		VulkanFence(const VkDevice& _logicalDevice, const uint8 _count = 1);
+		VulkanFence(const VkDevice& _logicalDevice, const uint16_t _count = 1);
 		~VulkanFence();
 
-		void Wait(const uint8 _fenceIndex = 1) const;
-		void Reset(const uint8 _fenceIndex = 1) const;
+		void Wait(const uint16_t _fenceIndex = 1) const;
+		void Reset(const uint16_t _fenceIndex = 1) const;
 		std::vector<VkFence>& Get() { return m_Fences; }
 
 	private:
 		VkDevice m_LogicalDevice;
 		std::vector<VkFence> m_Fences;
 	};
-} // End of Banshee namespace
+}
