+++ BansheeEngine/BansheeEngine/source/graphics/Window.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "foundation/Platform.h"
+#include "Foundation/Platform.h"
 
 struct GLFWwindow;
 
@@ -9,17 +9,16 @@ namespace Banshee
 	class Window
 	{
 	public:
-		Window(const uint16 _width, const uint16 _height, const char* _title);
+		Window(const uint16_t _width, const uint16_t _height, const char* _title);
 		~Window();
 
 		bool ShouldWindowClose() const;
 		void Update();
 		GLFWwindow* GetWindow() const { return m_Window; }
-		uint32 GetWidth() const;
-		uint32 GetHeight() const;
+		uint16_t GetWidth() const;
+		uint16_t GetHeight() const;
 
 	private:
 		GLFWwindow* m_Window;
 	};
-
-} // End of Banshee namespace
+}
